packages/ipc_router/pyproject.toml
[project]
name = "ipc-router"
version = "0.1.0"
description = "Core routing service for AegisIPC framework"
readme = "README.md"
requires-python = ">=3.13"
license = { text = "MIT" }
dependencies = [
    "nats-py==2.10.0",
    "msgpack==1.1.0",
    "pydantic==2.10.4",
    "pydantic-settings==2.7.0",
    "asyncio==3.4.3",
    "uvloop==0.21.0",
    "fastapi==0.115.6",
    "prometheus-client==0.21.0",
    "ipc-client-sdk",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.4",
    "pytest-asyncio==0.25.0",
    "pytest-cov==6.0.0",
    "pytest-mock==3.14.0",
    "httpx==0.28.1",
    "ipc-client-sdk",
]

[tool.uv.sources]
ipc-client-sdk = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipc-client-sdk",
]

[tool.hatch.build.targets.wheel]
packages = ["ipc_router"]
